Initial changes
Moved required functions to emailListSlice and removed obsolete functions from metaSlice, including metaSlice itself.
