+++ src/components/EmailDetail/DetailNavigation.tsx
@@ -5,13 +5,12 @@ import CircularProgress from '@mui/material/CircularProgress'
 import { selectLabelIds, selectStorageLabels } from '../../Store/labelsSlice'
 import * as S from './DetailNavigationStyles'
 import { selectCurrentEmail, selectViewIndex } from '../../Store/emailDetailSlice'
-import { loadEmails } from '../../Store/metaListSlice'
 import CloseMail from '../../utils/closeEmail'
 import NavigateNextMail from '../../utils/navigateNextEmail'
 import NavigatePreviousMail from '../../utils/navigatePreviousEmail'
 import { CustomIconLink } from '../Elements/Buttons'
 import loadNextPage from '../../utils/loadNextPage'
-import { selectEmailList } from '../../Store/emailListSlice'
+import { loadEmails, selectEmailList } from '../../Store/emailListSlice'
 import { selectIsLoading, selectIsSilentLoading } from '../../Store/utilsSlice'
 import { loadDraftList, selectDraftListLoaded } from '../../Store/draftsSlice'
 import * as draft from '../../constants/draftConstants'
