+++ src/components/EmailList/EmailList.tsx
@@ -2,9 +2,8 @@ import React, { useEffect, useMemo } from 'react'
 import CircularProgress from '@mui/material/CircularProgress'
 import { useLocation } from 'react-router-dom'
 import EmailListItem from '../EmailListItem/EmailListItem'
-import { loadEmails, refreshEmailFeed } from '../../Store/metaListSlice'
 import { loadDraftList } from '../../Store/draftsSlice'
-import { selectEmailList } from '../../Store/emailListSlice'
+import { loadEmails, refreshEmailFeed, selectEmailList } from '../../Store/emailListSlice'
 import { selectLabelIds, selectLoadedInbox } from '../../Store/labelsSlice'
 import { selectIsLoading } from '../../Store/utilsSlice'
 import Emptystate from '../Elements/EmptyState'
