+++ src/components/EmailOptions/ThrashMail.ts
@@ -1,4 +1,4 @@
-import { UpdateMetaListLabel } from '../../Store/metaListSlice'
+import { UpdateEmailListLabel } from '../../Store/emailListSlice'
 import { LocationObjectType } from '../types/globalTypes'
 
 interface ThrashMailProps {
@@ -14,7 +14,7 @@ const ThrashMail = (props: ThrashMailProps) => {
 
   const MarkEmailThrashed = () => {
     dispatch(
-      UpdateMetaListLabel({
+      UpdateEmailListLabel({
         messageId,
         request,
         location,
