+++ src/components/Inbox/InboxRefreshOption.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react'
 import { MdRefresh } from 'react-icons/md'
+import { refreshEmailFeed, selectIsFetching } from '../../Store/emailListSlice'
 import { useAppDispatch, useAppSelector } from '../../Store/hooks'
 import { selectLabelIds } from '../../Store/labelsSlice'
-import { refreshEmailFeed, selectIsFetching } from '../../Store/metaListSlice'
 import { selectIsLoading } from '../../Store/utilsSlice'
 import { CustomIconLink } from '../Elements/Buttons'
